Refactor imports with sed
