+++ lighthouse-plugin-publisher-ads/audits/ad-request-from-page-start.js
@@ -14,8 +14,9 @@
 
 const NetworkRecords = require('lighthouse/lighthouse-core/computed/network-records');
 const {auditNotApplicable} = require('../utils/builder');
-const {AUDITS, NOT_APPLICABLE, WARNINGS, formatMessage} = require('../messages/messages.js');
+const {AUDITS, NOT_APPLICABLE, WARNINGS} = require('../messages/messages');
 const {Audit} = require('lighthouse');
+const {formatMessage} = require('../messages/format');
 const {getAdStartTime, getPageResponseTime, getPageStartTime} = require('../utils/network-timing');
 
 const id = 'ad-request-from-page-start';
