+++ lighthouse-plugin-publisher-ads/audits/duplicate-tags.js
@@ -14,9 +14,10 @@
 
 const NetworkRecords = require('lighthouse/lighthouse-core/computed/network-records');
 const {auditNotApplicable} = require('../utils/builder');
-const {AUDITS, NOT_APPLICABLE, formatMessage} = require('../messages/messages.js');
+const {AUDITS, NOT_APPLICABLE} = require('../messages/messages');
 const {Audit} = require('lighthouse');
 const {containsAnySubstring} = require('../utils/resource-classification');
+const {formatMessage} = require('../messages/format');
 const {URL} = require('url');
 
 const id = 'duplicate-tags';
