+++ lighthouse-plugin-publisher-ads/audits/first-ad-paint.js
@@ -13,8 +13,9 @@
 // limitations under the License.
 
 const {auditNotApplicable} = require('../utils/builder');
-const {AUDITS, NOT_APPLICABLE, WARNINGS, formatMessage} = require('../messages/messages.js');
+const {AUDITS, NOT_APPLICABLE, WARNINGS} = require('../messages/messages');
 const {Audit} = require('lighthouse');
+const {formatMessage} = require('../messages/format');
 const {isGptIframe} = require('../utils/resource-classification');
 
 const id = 'first-ad-paint';
